# Integration of AES-XTS AArch64 implementation