--- a/crypto/fipsmodule/aes/mode_wrappers.c
+++ b/crypto/fipsmodule/aes/mode_wrappers.c
@@ -127,13 +127,19 @@ void AES_cfb128_encrypt(const uint8_t *in, uint8_t *out, size_t length,
 }
 
 #if defined(HWAES_XTS)
-void aes_hw_xts_cipher(const uint8_t *in, uint8_t *out, size_t length,
+int aes_hw_xts_cipher(const uint8_t *in, uint8_t *out, size_t length,
                        const AES_KEY *key1, const AES_KEY *key2,
                        const uint8_t iv[16], int enc) {
+  // The assembly functions abort on the following condition.
+  // They can be modified to return 0/1 instead of void, but
+  // this is the easy way out for now.
+  if (length < 16) return 0;
+
   if (enc) {
     aes_hw_xts_encrypt(in, out, length, key1, key2, iv);
   } else {
     aes_hw_xts_decrypt(in, out, length, key1, key2, iv);
   }
+  return 1;
 }
 #endif