--- a/decrepit/xts/xts.c
+++ b/decrepit/xts/xts.c
@@ -232,13 +232,8 @@ static int aes_xts_cipher(EVP_CIPHER_CTX *ctx, uint8_t *out,
 
 #if defined(HWAES_XTS)
   if (hwaes_capable()) {
-    if (ctx->encrypt) {
-      aes_hw_xts_encrypt(in, out, len, xctx->xts.key1, xctx->xts.key2,
-        ctx->iv);
-    } else {
-      aes_hw_xts_decrypt(in, out, len, xctx->xts.key1, xctx->xts.key2,
-        ctx->iv);
-    }
+      aes_hw_xts_cipher(in, out, len, xctx->xts.key1, xctx->xts.key2,
+                         ctx->iv, ctx->encrypt);
   } else {
 #endif
     if (!CRYPTO_xts128_encrypt(&xctx->xts, ctx->iv, in, out, len, ctx->encrypt)) {