--- a/include/openssl/cipher.h
+++ b/include/openssl/cipher.h
@@ -643,6 +643,12 @@ BSSL_NAMESPACE_END
 
 #endif
 
+/*
+ * The maximum permitted number of cipher blocks per data unit in XTS mode.
+ * Reference IEEE Std 1619-2018.
+ */
+#define XTS_MAX_BLOCKS_PER_DATA_UNIT            (1<<20)
+
 #define CIPHER_R_AES_KEY_SETUP_FAILED 100
 #define CIPHER_R_BAD_DECRYPT 101
 #define CIPHER_R_BAD_KEY_LENGTH 102
@@ -669,5 +675,7 @@ BSSL_NAMESPACE_END
 #define CIPHER_R_WRONG_FINAL_BLOCK_LENGTH 123
 #define CIPHER_R_NO_DIRECTION_SET 124
 #define CIPHER_R_INVALID_NONCE 125
+#define CIPHER_R_XTS_DUPLICATED_KEYS 138
+#define CIPHER_R_XTS_DATA_UNIT_IS_TOO_LARGE 139
 
 #endif  // OPENSSL_HEADER_CIPHER_H