### Issues:
Resolves CryptoAlg-929

### Description of changes:
This PR integrates the AArch64 implementation of AES-XTS into AWS-LC. The source is from this OpenSSL's commit: 
https://github.com/openssl/openssl/commit/9ce8e0d17e608de4f85f7543c52b146e3c6a2291:
* Some comments were added to the Perl Assembly implementation on `{prefix}_xts_encrypt`.
* BTI instructions were added at the beginning of `{prefix}_xts_encrypt` and `{prefix}_xts_decrypt`.
* Test vectors were added in order to test the different paths of the AArch64 implementation.
* Checks that the 2 keys are different and that the input size does not exceed 2^{20} are also added to the functions `aes_xts_init_key` and `aes_xts_cipher`, respectively.

### Call-outs
* `libdecrepit` was added as a dependency of the `bssl` tool in order to run the speed test.
* The assembly code added should work on Arm (other than Armv8) and would be chosen on those platforms but it was not tested other than on Armv8.

### Performance improvement
#### Init and Encrypt
|      Size   | Before (MB/s)(1) | After (MB/s) | Improvement |
--------------:|------------------:|--------------:|:--------------:
|    16 bytes |        88        |       91     |    1.03x    |
|   256 bytes |       647        |      883     |    1.36x    |
|  1350 bytes |       971        |     1804     |    1.86x    |
|  8192 bytes |      1097        |     2334     |    2.13x    |
| 16384 bytes |      1109        |     2398     |    2.16x    |

#### Init and Decrypt:
|      Size   | Before (MB/s)(1) | After (MB/s) | Improvement |
--------------:|------------------:|--------------:|:--------------:
|    16 bytes |        83        |       85     |    1.02x    |
|   256 bytes |       621        |      810     |    1.30x    |
|  1350 bytes |       948        |     1746     |    1.84x    |
|  8192 bytes |      1091        |     2289     |    2.10x    |
| 16384 bytes |      1105        |     2356     |    2.13x    |

(1) C implementation which includes AES block encryption in assembly.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.